+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add codecov
 * Add tidiers for LDA objects from topicmodels
 * Fixed a bug when tidying by line/sentence/paragraph/regex and there are multiple non-text columns
+* Added ability to pass a (custom tokenizing) function to token. Also added a collapse argument that makes the choice whether to combine lines before tokenizing explicit.
 
 # tidytext 0.1.0